src/api/middlewares/logging_middleware.py
import time
from fastapi import Request
from starlette.middleware.base import BaseHTTPMiddleware


class LoggingMiddleware(BaseHTTPMiddleware):
    """
    Middleware для логирования всех запросов
    """

    async def dispatch(self, request: Request, call_next):
        start_time = time.time()

        response = await call_next(request)

        process_time = (time.time() - start_time) * 1000
        print(f"[LOG] {request.method} {request.url.path} - {response.status_code} - {process_time:.2f}ms")

        return response
