requirements.txt
sqlalchemy
pydantic
jwt
bcrypt 
fastapi 
uvicorn 
asyncpg
python-dotenv 
aiosqlite